New constructors are added to parse through the DesignCompositeManager
This dependency injection chain was added because externally it is not
possible to obtain the DesignComposite before it is created. However we
can grab the DesignPage at creation and pass in a DesignCompositeManager
object. This will also allow us to hide/show the various other toolbars
on the Properties view as well.