Specify a specific jar name to be used as icon pack
Instead of listing all the resources in the available classpaths
A specific jar name can now be specified. This will be the only jar
displayed in the classpath list of the icon image chooser. This was
added to avoid clients using the wrong classpaths by limiting their
options to only that of a specific icon pack jar