Adjusting crop method to prevent negative height and width values
Having a negative width or height value of the Rectangle, produces an
error and makes no sense. This negative value occurs when a border has
been set however the rectangle has only a height or width of 1. If this
is the case then the crop method tries to reduce the width/height of the
rectangle by the width/height of the Insets. This can potentially reduce
the height/width of the rectangle to a negative value and this leads to
an error.

Updated contributor documentation

Removed auto added Override annotations