Adding a target file
This target file located in target-platform/wb.target can be used to
configure your IDE with the necessary libraries. Simply open the file
and press "Activate" to configure your IDE the same way as the Tycho
build is configured. This will also allow us to move certain libraries
from being copied into the project to Maven dependencies.

For #187