Update SwtTests test suite
This disables / removes any failing test from the test suite.
From the original 889 test we still have 869. The next step would be add
these tests to the Tycho build to avoid regressions in the future but
with this change you can at least run the test suite from the IDE
without any issue.

Search for DISABLE_ to find the test which have been deactivated.

Fixes #239