Improve the target platform setup for better reuse
Split out from wb.target, the Maven locations into a separate a
wb-maven.target that can be included/composed into both the wb.target
and in the targlet task of the WindowBuilder.setup.

Tycho 2.7.3 can consume such a composed *.target and the targlet setup
avoids duplicating the resolution of the non-Maven content to the target
platform.