Fixes tests after update of Win11
The adjusted tests contain hard coded values for the size of dialogs,
etc. and win11 seem to have changed these values. To make it possible to
run these tests without issue on Win10 and Win11, the value were just
lowered to the reported values from Win11. While not the perfect
solution a rework to use the correct values on each operation system is
beyond my desired time investment.

Fixes #286