Ensures that org.eclipse.wb.swt.ResourceManager is available for the ResourceManagerTest
The ResourceManagerTest assumed that the first test in the file is
executed as a first test as this would ensure that the ResourceManager
class is available in the test.

The waitForAutoBuild() checked for such cases but it was hard to see. So
by commenting assertEquals() at the end of this method out and by adding
he following code after the waitForAutoBuild() call in the test I could
use the runtime IDE after the failing situation to see what actually is
the case in Eclipse.

	Display.getDefault().syncExec(() ->

	{
		while (true) {
			Display.getDefault().readAndDispatch();
		}

	});

This way it was easy to see that the corresponding file was not present
the generated test code.