Upload test results, even on test failures.
Currently, the build job is aborted when at least one test has failed.
As a result, none of the test results are uploaded to GitHub, and the
developer has to either manually dig through the build log or retest
them manually. With this change, the Maven build will ignore test
failures and the pipeline instead fails after evaluating the test
results.