Repair & Execute Nebula Tests
The Nebula tests attempt to fetch the Nebula jars from a non-existent
plugin. Afaik, this plugin has never been under version control.

Instead, add the Nebula update sites to the target platform and load the
plugins from there.

The CTableTree tests have been disabled, as this widget has been
archived back in 2009 and is not included in the Nebula releases,
meaning there is no way to test it. See Bug 162138 for reference.