The artifact name for the test results should be unique.
The property 'matrix.config.native' is not set on the build nodes.
Therefore, the artifact name is e.g. 'test-results--java11' for the Java
11 builds, for both the Windows and Linux build. This means the that the
test results that are uploaded first are overwritten by the ones that
are uploaded last, meaning we only upload half the test results and miss
out on potential test failures!

By using 'matrix.os', the same property used in the job matrix, we can
ensure that it exists and - in combination with the Java version - is
unique.