Remove GTK2 from Linux-based OS Support
Remove GTK2 code from OSSupportLinux, as well as the binaries and C
code. Instead, GTK3 (or higher) is assumed by default.

This code is causing problems on newer systems using GTK3. As an
example, the rectangle containing the toolBarManager and menuManager are
offset by the height of the title bar, while the widgets are not.

Instead of trying to remove the offset of the figures, it seems more
reasonable to restore the title bar, in order to match the
representation on other operation systems.

It seems like the title bar was disabled due to incompatibilities with
GTK2 and compiz. It seems like this problem no longer exists with
GTK3.

Tested with KDE Plasma