Fail verification build if we see issue
The setup introduced with 3b9a81b
results in the false impression that everything works fine, even with
our currently lots of unit test errors.

I suggest we have the verification build again if we have test errors so
that we have to fix or disable the failing tests and make it easier for
use to avoid changes which result in new failing test.