Fail the verification build if we have test errors
The setup introduced with 3b9a81b
results in the false impression that everything works fine, even with
our currently lots of unit test errors.

I suggest we have the verification build again if we have test errors so
that we have to fix or disable the failing tests and make it easier for
use to avoid changes which result in new failing test.

Test results are uploaded, regardless of test failures, so that they can
be accessed by the "Publish Test Results" action. The fast-fail has been
disabled, to ensure that test failures in one matrix job doesn't cancel
the remaining, still running jobs. Lastly, the "Publish Test Results"
action has been adapted to not only publish the results on success, but
also on failure.