Adapt platform-specific checks to both work on Windows and Linux.
The size and position of widgets may differ on Windows and Linux. Any
test that uses pixels to calculate the location of a figure has been
adapted to work on both systems.

Because the tests are run in the UI thread, it is sometimes necessary to
explicitly wait for the execution of other UI jobs, such as redrawing
widgets. Examples are tests that select widgets
based on their position in the design page.