Fix calculation of shell insets in CoordinateUtilsTest
While the shell is invisible, its coordinates are not properly updated.
When attempting to calculate the shell location and client area insets,
GTK returns junk coordinates to the WB. Which then results in
nonsensical insets and bounds.

Note that the test has currently been disabled to due failure on the
build nodes. The title bar doesn't seem to have a height. See #389.