[GTK3] Fix freeze when using GTKLookAndFeel in Swing
Linux synchronizes GTK calls. Meaning if this method is called from
within the SWT UI thread, then we may have already acquired the lock. If
we then execute the job in the AWT UI thread and wait for its
completion, we risk a deadlock, as the AWT thread may also try to
acquire the same lock.

In order to avoid this problem, the SWT thread has to explicitly leave
the critical region before and re-enter it, immediately after the AWT
job has been executed. This operation is safe, as the current thread
blocks any further SWT updates, meaning the AWT thread is the only one
who can interact with GDK for this brief duration.

Note that this behavior is only relevant for GTK3. Those methods have
been marked as deprecated in GTK 3.6 and removed in GTK4. Operations are
assumed to always be executed in the main thread, rendering this problem
obsolete.