Fix the remaining MigLayoutConverter test case.
Getting the button to be aligned with the center requires pixel-perfect
precision. The current width of the button is too small and due to
rounding errors, is always aligned to the left.

Increase the button width to add a little bit more tolerance so that it
works on all platforms.