[Oomph] Add missing Nebula repositories to the setup script
With the addition of the Nebula repositories, the Oomph setup script is
no longer working due to unresolved dependencies.

Add the update sites containing the nightly builds to the script, in
order to detect API incompatibilties as soon as possible.

Note that the Tycho build is using a fixed release, instead of the
nightly build, simply to make sure that the build results are
reproducible.