Adapt test cases to the removal of the Resource Navigator.
This view has been marked as deprecated in 3.5 with the recommendation
to use the Common Navigator Framework instead.

As part of this removal, the ID_RES_NAV page layout is no longer going
to be available, which is referenced in two of our test cases. Use the
Project Explorer instead, given that it provides almost the same
functionality, which is more than sufficient for the tests.