Extend unstable tests with additional checks to isolate point-of-failure
Create a new assertJavaInfo method, to check the whether a widget with
the given name exists in the Java code. This is the underlying cause of
the Jenkins failure, because no matching data binding can be found.

Reactivate test_deleteItemWithSubmenu() as this is a general problem and
not limited to just this test.