[JUnit] Pump the SWT event queue for 100ms after selecting a tree object
The problem is that the property table is redrawn while WB is still
deleting the widget. So we're at a state where the widget has already
been removed from the databinding but is still registered as selected.

When we then try to redraw the table, we attempt to fetch the
no-longer-existing databinding for the widget, causing a model
exception.

In short, instead of:

- Update Selection
- Update Databinding
- Redraw

We do:

- Update Selection
- Update Databinding
- Redraw <- Error

We already tried something similar with revision
ea5c81c078bb654efbe6214610e960b6dc1b0502. However just doing a single
SWT event loop isn't enough.

Resolves #413