Completely remove the dependency to Commons Primitives
This dormant project hasn't been updated since 2003 and is only used at
two locations in WB, to handle the auto-boxing of primitive types.

There is a minimally more complex expression, with almost the same
effect. I think it's more elegant to replace those two references and
get rid of this dependency completely, rather than to continue carrying
it around.

As a side-note, the toArray() method in the ArrayIntList internally
constructs a new array anyway. So even without it, we're still at the
same level of complexity, i.e. O(n).