[GEF] Replace Translatable interface with its latest GEF component
As a "major" change, the translate() methods have been renamed to
performTranslate().

Furthermore, a new performScale() method has been introduced, which
needs to be implemented by the classes implementing this interface.

For the sake of simplicity, the Point class extends the GEF Point. This
provides compatibility with the corresponding performTranslate method of
the GEF interface while at the same time minimizes the amount of code
modification that has to be done.