Fix: "lastgoodbuild" update site misses dependent bundles
Because we started replacing embedded jars with their latest version
from Maven Central or the SimRel repository, it is no longer possible to
install WindowBuilder in a fresh Eclipse installation due to unresolved
dependencies.

While it is possible to resolve this by manually adding those update
sites, this is highly inconvenient and - for Maven dependencies - not
always possible.

Add those plugin to the relevant WB feature, to ensure that no left-over
plugin remains, in case the feature is removed from the Eclipse
installation.