Issue 352 - Parameterized JComboBox breaks Design tab
The problem boils down to the fact that generic type information are not
stored inside the Java byte code. Meaning a generic method 'foo(T t)' is
stored as 'foo(Object t)'.

In principle, WindowBuilder is able to correctly map those methods. But
only for normal objects, not for arrays. This oversight has been
corrected.

The reason it showed up for the JComboBox specifically is because of
its generic constructor '<init>(T[] items)', which is stored as
'<init>(Object[] items)'.

Because our mapping mechanism wasn't working for arrays, we try to look
up the constructor '<init>(String[] items)' when trying to instantiate
the object. Because no match was found, an exception is thrown.