[GEF] Replace Rectangle with latest GEF counterpart.
The side-effects from setX() and setY() have been removed. Meaning that
updating the x value also updates the width of the rectangle. And
similarly the y value and height. Instead, the height has to be
explicitly updated.

The methods moveX() and moveY() have been renamed to shrinkLeft() and
shrinkTop().

Instead of passing the direction to the getInterval() method, there are
now two new methods getHorizontalInterval() and getVerticalInterval(),
to get the horizontal and vertical interval, respectively.