[GEF] Replace Interval class with latest GEF class.
Due to the GEF class being an immutable record, all transformation
return a new instance instead of modifying the current instance. This
means we have to update the local variable with the new value, whenever
we call growLeading() or growTrailing().

Furthermore, the local fields are now private. Access has to be done via
the getter methods. This affects both 'begin' and 'length'.

The method getRightmostIntervalIndex() has been renamed to
getRightMostIntervalIndex().