Replace CGLib code generation in WB Core Java with ByteBuddy.
The DefaultMethodInterceptor extends the InvocationHandler. It has been
adjusted to only handle abstract methods, for which a default value is
used. In order to invoke the parent method, SuperMethodCall.INSTANCE may
be used.

The callback argument has been removed from the
createAnonymousInstance() method, as it is always the
DefaultMethodInterceptor.

For the ExecutionFlowUtils, the enhancer has been adjusted to only
generate a single enhanced class is used for each ASTVisitor. Reason
being that the ExecutionFlowUtils classloader is used by ByteBuddy.
Given that this classloader is never destroyed, each enhanced class is
kept in memory indefinitely. If left unchecked, this may lead to an
OutOfMemoryError. For each instance of this class, the currently active
VisitingContext, ExecutionFlowDescription and ExecutionFlowFrameVisitor
is stored in a local variable and used internally.

For the InvocationEvaluator, a type-cache has been introduced, to store
enhanced class for a given EvaluationContext. Each enhanced class is
unmodifiable and may be reused.

For the ThisCreationSupport, the method invocation is delegated to a
non-static field of the enhanced object. This field is set after an
instance has been created. In case enhanced methods are invoked within
the constructor, meaning this field hasn't been initialized yet, the
referee is fetched from a thread-local cache. This cache is set
immediately before, and cleared immediately after the object has been
created.

In short, following issues have been observed when migrating from CGLib
to ByteBuddy:

- Due to high cost of loading a class, the enhanced classes should be
reused as much as possible..
- The enhanced class should be stored in a cache. Instances of this
class should be created from the same class object, rather than creating
a new class object every time.
- The InvocationHandler should be used as much as possible, in order to
keep the number of required packages for the class loaders as low as
possible.

The reason for those big changes are due to the fact, that CGLib is
using a cache, to avoid having to load the same class multiple times.
ByteBuddy doesn't use such an implicit cache, because the enhanced
classes are permanently stored in memory and thus create a memory leak.

Not using a cache has a major performance impact. Most noticeably, our
tests take twice as long to finish. Which means we explicitly have to
create a manage such a cache ourselves.