Enhanced classes should be directly injected in the parent class-loader
Using this strategy does not create a new class loader but injects all
classes into the given ClassLoader by reflective access. This prevents
the loading of classes with cyclic load-time dependencies but avoids the
creation of an additional class loader.

Note that this strategy usually yields a better runtime performance.