Specify a global timeout of 90 min. for the GitHub build jobs (#471)
The Jenkins build currently sets two timeouts. A 90 min. timeout for the
entire build and a 60 min. timeout for the JUnit tests.

Whenever we want to adjust the timeout, we have to separate places to
edit (the pom.xml file on the test bundle and the Jenkinsfile). We
should drop the timeout of the tests and simply rely on the global
timeout.

Note that we GitHub workflow already has a timeout of 5h by default. It
is lowered to 90min to avoid excessive build times in case it ever gets
stuck and to keep in in sync with the Jenkins build.