Validate the Manifest version numbers using the latest release.
In order to reduce the chance of accidentally increasing the version
number of a plug-in after it has been modified, we should use the
tycho-baseline-plugin to automate this task.

The versions are compared against the latest release. If the versions
are identical (excluding the qualifier) or if the increment is
insufficient, an error is thrown.

The commit date is used to create reproducible timestamps. Because this
means that the build date of the latest release happened after they are
commited, we now have to increment all Manifest versions, in order to
prevent the timestamp from going backwards.