Update Code generation to the new Observable Binding API - #491
The deprecated class SwtObservables & friends have been removed with the
Eclipse 4.26. Our code-generation for creating bindings is still
referencing those classes, thus leading to an internal error when using
them. The code generation has been adapted to exclusively use the new
API.

The JUnit tests have been reactivated and repaired. Where necessary, the
parser has been extended to support as much of the new API as required
by the tests.

Furthermore, additional utility classes for Java 8 or higher need to be
created. Those are currently just a copy of the Java 7 classes with no
further modifications.