[GEF] Replace Graphics class with its GEF counterpart.
The getTextExtent() method in the Graphics class has been replaced by
its equivalent method in TextUtilities. Where necessary, an
instance of Graphics is created via the SWTGraphics class.

The low-level calls to the GC instance are almost identical between both
implementation. The only noticeable difference is the internal state of
the Graphics class, but which doesn't seem to cause any major UI
changes.