Update to Tycho 5 nightly build.
Tycho 4 has been released last week, meaning that the 4.0.0 nightly
builds are no longer available. Instead of using the 4.0.0 release, we
should continue to use the nightly builds, in order to respond to
incompatibilities as fast as possible.

The pom-less extension has to be set to the latest release, as the
snapshots are not available.

Note that even though Tycho 5 is still using Maven 3.9 right now, it
will eventually require Maven 4 in the future.