Fix "Widget is Disposed" exception in FontDialogField.
When updating the font label, WindowBuilder creates a new font instance.
This instance is never disposed properly, leading to the mentioned
exception.

Given that the dialog field doesn't have a dispose() method, we can't
use a resource manager to track this object. Instead, the lifecycle is
bound to the lifecycle of the label it is assigned to.