[GEF] Figure should implement the IFigure interface.
Most of the GEF/Draw2D API expects an object of type IFigure as an
argument. In order to simplify the migration to downstream GEF, our own
figure implementation should implement this interface so that we can
gradually build back our own classes.

Instead of implementing the entire IFigure interface, we instead extend
the Figure base class, overwriting all non-trivial (and non-final)
methods with the already existing implementation. Due to a
bi-directional relation between Figure and Border, it is also necessary
for our Border base-class to implement the corresponding Draw2D
interface.