[GEF] Replace IFigureListener with FigureListener
The upstream FigureListener class doesn't contain the figureReparent()
method. Instead, a property change event with name "parent" is fired.
The figureReparent() method is only used by the AncestorEventTable. To
ensure compatibility, this class now also implements a
PropertyChangeListener which responds to this event, whenever the parent
of a figure is updated.