Remove singleton pattern from ObjectsLabelProvider
In order to switch to ImageDescriptors for IObjectPresentation, we first
need to remove the singleton pattern from this label provider.
We can create images from those descriptors using a resource manager.
But if the instance of this provider is static, then those instances are
not properly disposed.

The getImage() and getText() methods have been moved to ObjectInfo.