Set "force" flag so that test project files are in sync with file system
Some of our tests occasionally fail with error messages of the form
"Resource '/TestProject/src/test' does not exist." It seems like some of
the cleanup operations between the test executions cause the workspace
to become out-of-sync with the file system. Instead of forcing a
synchronization, our tests then simply throw an exception while still
leaving everything in an inconsistent state.
The "force" flag has been set for the resource creation, to hopefully
avoid errors like this in the future.