Add MVEL bundle to the project classloader
This resolves the NoClassDefFoundError thrown by MVEL when trying to
parse an expression via ScriptUtils. Internally, MVEL loads a class
using our composite classloader which itself is based on the project
classloader. This classloader is not the one that called MVEL and
doesn't have its classes in its classpath. Simply adding this missing
dependency solves this problem.