Ensure that DisposeProjectAfter is always called after the test
If the test happens to fail due to an internal error, the cleanup task
is never executed. This is resolved by puttin it in the finally block.
Futhermore, we wait 10ms before and after disposing the project to
ensure that the Eclipse had enough time to properly delete the project..