[GEF] Replace RefreshManager with UpdateManager
This is the first step of migrating to the LightweightSystem and
ViewPort structure of the GEF FigureCanvas. An instance of the manager
is normally created by the LightweightSystem and accessible by every
figure via the getUpdateManager() method. Because we don't use this
mechanism yet, we have to explicitly create and return it in the root
figure.