[GEF] Simplify proxy event handling in EventManager
We currently use a complex implementation in order to support a
"delayed" execution of SWT events by intercepting the listener
invocations via a proxy class.
If we instead route the relevant invocations through a common method and
wrap the listener invocation around a runnable, then we can achieve the
same effect.
This step is a pre-requisite for synchronizing our EventManager
implementation with the GEF EventDispatcher in the future