Fix: Palette not properly updated when hovering over items
The item beneath the cursor is not drawn as etched-in. This is because
this area is not included in the "damaged" area and thus not repainted.
Adjust the calculation of the dirty region to match the code in the
deleted RefreshManager to fix this issue.

Amends 7e95ab22602b957fb15106bbfb62c0b8df3d8b95