[GEF] Replace EventTable with EventListenerList
This ensures that methods called by e.g. Figure.handleMousePressed()
notify all listeners that have been registered by the figure.

Because we use two event tables, one for WindowBuilder and one for GEF,
it might happen that listeners were skipped, when calling the methods
inherited from the parent figure. This change merges the WindowBuilder
table into the GEF table.

Unlike our implementation, the GEF event table allows the same listener
to be added twice and will throw an error, when null is added.