Fix: WindowBuilder can't be installed in Eclipse IDE
The Eclipse IDE doesn't depend on commons-collections. Because we have a
direct dependency to it, this prevents WindowBuilder from being
installed without additional update sites. Furthermore, this breaks the
dropin mechanism.
The solution consists of simply adding it to the dependencies feature so
that this dependency is also included in the update site.