Fix: Occasional NPE when creating a mocked widget
It may happen that inside the constructor, we create and run a separate
thread. Because the method interceptor is stored in a thread-local
variable, it is not available to this thread.
Instead, the interceptor is stored in an atomic variable. Note that this
requires that the object creation has to be synchronized. Otherwise it
might happen that an instance created in one thread is using the
interceptor of another instance created in a different thread.