Second attempt to stabilize the PaletteUITests
Adjust the chooseComponentClass() to wait for the search text to be
updated. It seems like the query from an earlier test is carried over.
Furthermore, adjust the waitFor() method to always execute at least one
cycle, in order to run idle UI events.

Revert "Fix: Deadlock when an exception is thrown inside an UIRunnable."
This reverts commit 9d88d576a74534c300b41561f94b960647115f41.