[GEF] Synchronize EventManager with GEF EventDispatcher
In order to migrate to the GEF FigureCanvas, we need to first implement
a compatible EventDispatcher. The event-listener interfaces are no
longer implemented by the event dispatcher, but rather by the
lightweight system. Given that we don't use it yet, this class is
imitated by adding the listeners directly to the canvas.