Fix: Potential memory leak due to undisposed resource registries
When we create instances of various ResourceRegistrys, for example
ColorRegistry, they usually want to be disposed with Display, so they
use Display#disposeExec(Runnable). But our Display instance lives all
time while Eclipse lives. So, practically we have memory leak: we keep
in memory instance of ResourceRegistry. its allocated resources, and
what is much worse - instance of ClassLoader that loaded this
ReourceRegistry.

The old implementation that used to explicitly dispose those objects was
no longer functional and has been repaired.