[GEF] Replace ScrollModel classes with RangeModel and ViewPort
Our ScrollModel class handles two tasks: It manages the start, and and
extent of the horizontal/vertical scrollbar as well as updating the SWT
widget. In downstream GEF, this has been split up into two separate
classes. The RangeModel contains the possible values of the scrollbar
and the ViewPort updates the ScrollBar widget.

To be able to use the ViewPort, our FigureCanvas implementation needs to
extend the GEF FigureCanvas. As an additional benefit, this then also
allows us to directly inject our custom EventManager for handling
Mouse/Keyboard events.