Move LocalResourceManager creation to separate method
For the sake of simplicity, the LocalResourceManager instance should be
created inside a separate method. Doing so makes it much easier to
ensure that the object is created as early as possible.

More specifically, after a call to super(), the root info created (e.g.
via new Shell()) and before its used inside another widget.